+++ test_canvas_web_drop_v10121.py
@@ -21,7 +21,7 @@ class Root:
 
 class CanvasWebDropV10121Tests(unittest.TestCase):
     def test_version(self):
-        self.assertEqual(APP_VERSION,'1.0.145-rc16')
+        self.assertEqual(APP_VERSION,'1.0.145-rc17')
         self.assertEqual(FILE_VERSION,'1.0.145')
 
     def test_local_file_drop(self):